refactor: bookstore_v1.0.4.sql
book 테이블
book_view_count 타입변경
bigint -> int

+++ erd-cloud/bookstore_v1.0.4/bookstore_v1.0.4.sql
@@ -59,7 +59,7 @@ CREATE TABLE book (
 	is_packed BOOLEAN NOT NULL,
 	stock INT NOT NULL,
 	book_thumbnail_image_url TEXT NOT NULL,
-	book_view_count BIGINT NOT NULL DEFAULT 0,
+	book_view_count INT NOT NULL DEFAULT 0,
 	book_discount INT NULL,
 	PRIMARY KEY (book_id),
